Update README.md
Word and Character Counter 

By: Clayton Cosper 

This document will be an overview of the following program Word Counter. To start off lets overview what the purpose of this program is. This program is designed to take an input from the console window (this is a window in which the user can enter data into the program), such as a string of text or words. Once the user has entered the text into the console window the program will do three things. The first thing is to count the number of words by eliminating white space in between words and punctuation. The second set of directives will be to find the average word length. Last the program will find the largest word and tell you how many characters there are present in that word. 

The introduction to this program begins with section one which includes lines (1-8), The first 6 lines of the program are introducing three library’s of data to our program. When I talk about the library it is a collection of macros, types and functions for tasks such as input/output processing, string handling, memory management, mathematical computations and many other operating systems. There are three library's that make this program able to function correctly. Iostream allows us to use built in recognition code to read input from the keyboard. The second important library is a vector they are sequence containers that store elements. 

Now to overview the main part of the program. This is where all the action takes place. Section 2 begins with line 12 “int main()” this is the entry point to our code. Looking at the rest of section 2 it will include (lines 16-25). The next set of lines are our variable that will be used in our program. For example: line 17 “int char_count = 0” we will later use this to find the exact character count for the data that was entered into the console window. Within this example look at the letters “int” this is defining the variable to always be a number. 

 Looking at the last part of section 2 is where I introduce vectors and define that variable name called “passage”. Let's break down line 23 “vector<char>passage”. This line of code has three meaning to it. I am saying here I want to use vectors (are sequence containers that store elements) an you declare a vector by using these signs (<,>). Inside the vector I wrote “char” this means everything I want to be stored inside the vector will be a character. Last part of this line is where I declare the name for this variable. When working with vector you will always have another line of code when trying to save data from a secondary source such as the console window for example “char current” this line will play a big role in the next section. 

 

 

Section 3 will include subsections to better understand the processes that are happening in these lines of code. Section 3 will cover lines (29-65). The first thing we do is give the user or operator instructions on what they need to do. For example, look at line 29 (cout << "Please enter all your sentences below (hit enter when finished)” in this line of code the program is outputting information to the console window the user can see by using the “cout” statement. 

 Section 3.1 we look at a more complexed loop statement. This loop statement is on line 31 and says “while (cin.get(current))” this is used to read an unformatted string (set of characters) from the standard input device (keyboard) from the console window. This function reads a complete string until a given delimiter (one or more characters that separate text strings). In section 2 I talked about the importance of “char current”. After the user enters in the information it is stored as characters to the word current. Another way to understand this is think of a book you have your title name(variable) then inside the book you have pages of words (memory). Within one book you have a lot of stored text. Think of the word current as the book title and the user input as the pages. Then I take what is stored in the variable “current” and store it to our vector named passage. This is what this line of code is doing “passage.push_back(current)”. We store the string of text in the vector because it will store each character in its own box of memory rather than the entire text in one box of memory. 

Section 3.2 holds our switch statement that will comb through the users input for delimiter items, such as punctuation and spaces. Line 35 says “switch (passage[index])” this statement means we will be reading the first character in the vector. We start with the first character in the vector because index equals zero. Let's review vectors for a moment. When we store data inside a vector each character will have its own memory location. Any time you store data in memory it starts with the memory space [0]. For example: “The dog was happy.” Let's say we wanted to recall the number [6] the letter would be “g” or call upon memory space [2] the letter would be “e” and last let's look at memory space [0] that letter would be “T”. Looking at lines 37 and 38 these are to filter out punctuation. 

 

Section 3.3 we start off with a basic if loop. This is a type of loop that will continue to repeat until the condition inside of the paratheses becomes false. This if loop is on line 40. The first loop is false that means we jump down to line 58. This loop is false because at this current time word_length is still equal to zero. In this line I am saying to add 1 to the variable char_count. The next line I am saying to add 1 to word_length. This means both variables are now equal to 1 not zero anymore. This is because we have read the this first character in the program so now, we need to up our character count and word length. The program then falls to line 64 where I said “index++” this means index is now equal to one. This is because we have already checked the memory of the vector for slot [0] meaning we have already read the character for the first slot.  

The program then starts back over at line 32 because of the “while” loop. Now we will start over at the switch statement on line 35. Line 35 now says “switch (passage[index(1)])” index is now equal to one therefore the program begins to check the second character in the text the user entered. The program then checks the character against the switch statement. If no punctuation. Is present the program continues onto line 40. This statement is false because at this point wo process until all of the characters have been processed through the program. Since we upped our character count previously this statement is now true. We continue to our next if statement but first we must read line 42 that says word_count is now equal to one. Let's now test the if statement on line 43. Currently word_length is greater then max_length  because we added one to word length earlier. We then compute line 45 and declare that max_length is now equal to word_length. This is to find the longest word in the string(text) the user has entered. The last thing to do in is if loop is to re-assign word_length back to zero, you can see this happing in line 52. After the if statement we drop back down to our default statement and repeat the process until the entire document or text has been read. 

Finally, section 4 will have our output product back to the console window using the “cout” statement. The first line in section 4 will print out “"Number of words”. The second line will print “"Average word length” by dividing the following “char_count / word_count”. The last line of the program will print “Maximum word length” the user should now have been able to see everything the program was meant to do in the console window.