Create Money Counter
Money Calculator Program 
By Clayton Cosper: 

 

	This document will be an overview of the following program  Money Calculator. To start off lets overview what the purpose of this program is. This program is designed to take an input from the console window (this is a window in which the user can enter data into the program), and output a given answer based upon the mathematical equations that happen within the programs code. The Money Calculator was designed to take large amount of data and quickly turn that into a readable form in dollars or cents. The program lets you gain the knowledge of the exact value of your money that you currently have. 

	Starting at section one which will cover lines 7 and 8. These lines starts off the program with introducing the iostream library. This line of code is allowing us to pull the input and to deliver output to the console window. Without this line of code, we would not be able to receive any input that the user may enter into the program. Line 8 is a short cut for the program allowing us to skip the extra labeling on each line. Let's go ahead and move down to line 10. This line is declaring our functions in the later lines of code. 

 	Advancing to the next line of code is our variable declarations. Section three of our program will cover lines 13-23. We are declaring our variables that will be used in the program. At the same time, we are setting values to the variables by putting an equal sign on the other side of the variable name. Example look at line 13, we said float penny= .01. Float means there will be a decimal involved, penny is our variable name, and .01 is the value that penny will be worth. So, if we type into the console window we had 2 pennies in the console window the output would be .02. 

	Advancing to the next set of code it will cover section 4 lines 26 and 27. This line we have created another set of variables, but these sets of variables will be for our input purposes only. So once a user has entered in a value later in the program these variables will now be worth what the user has inputted. In the first section of variables lines 13-23, we already declared the worth of each item, but this second set of variables are unknown until they are declared later in the console window.  

The next overview will be of section 5 lines 32-204. This is the first line of code that will be executed on to the user’s screen. The words cout<< outputs information that the user can read. This line of code says, "Please enter the number of pennies in an amount greater than or equal to zero:". It gives the user or operator specific instructions to follow. Following that line of code, you will enter in the number of pennies the user has. This is when line 33 is executed, this line of code is invisible to the user but in the background the program will assign the value that the user inputs to the variable that is attached to the cin>>. The user cannot enter in any information unless there is a cin>> function in the program. Line 35 starts one of the most important parts of the program.  This line of code says while (penny_physical (This would be the value the user entered in the previously) is less than or equal to –1 ) execute what is in the squiggly brackets. So lines 35-40 will be executed repeatedly until the user enters in a number that is greater than or equal to zero. These lines of code will be executed over and over because of the ‘while’ loop we have before the conditional statement. This set of code is important because you cannot have a negative penny or negative dollar. The next set of commands will not execute until line 34 becomes false.  Once this line of code is false it skips down to the if statement on line 43. On line 43, We see another conditional statement. This statement is important because without it we could not continue onto the rest of the program. This statement says if penny_physical is greater than or equal to 0. We are testing here to see if the user has followed the directions and entered a correct integer. If this conditional statement is true, we execute the next lines of code in the squiggly brackets. Since this statement is inside of a ‘if’ condition we will only execute these sets of codes once.  Lines 34-205 are the same just with different variables being used. 

Section 6 will cover lines 208-229 this is where we retell the user or operator what they have entered so they can double check they inputted the correct information. Lines 208-230 will be printed to console window so that the operator can see the information they provided to the computer. If the user realizes they have inputted the wrong value for one of the variables. The user will have to start the program over.  

Moving to section 7 lines 232-243, this is where we have created equations using the variables in our program. Line 233 can be used as an example. This line says, “penny *= penny_physical” another way of saying this line would be the new value of penny is equal to penny times the number of pennies entered. You have to look at the variables as numbers and it will make more sense. Try this, penny = penny .01 x penny_physical (user entered 50). Lines 233 through 243 are looked at the same way. We are taking a variable and reassigning it a new value.  

The last section will cover lines 246-250.  The line starts off with a float total= this is because I made a new variable at the end of the program to capture the total amount of money you have. I made it a float variable because you will likely have some spare change.  After the equal sign I added each variable together to get the total amount of money.  I wanted to make line 246 more readable and user friendly. So, I created line 247, by separating the total of coins and single dollars from the total of larger dollar’s you can see more precisely what is being added together. Finally, the last line on the page is a message back to the console window letting the user know the total amount of money they have. With this last line of output this program has now successfully done its job with totaling the user's finances.